* исправлен баг с добавлением ролей * перемещена вся логика из контроллера в сервис * изменен тип данных ID с int -> Long * добавлен начальный контроллер профиля(/api/profile) * созданы модели Anime и CollectionOfficial * Создан Контролер AnimeController который может показывать Весь список аниме и по айди

+++ src/main/java/ru/Raingor/webAnimeSite/payload/response/JwtResponse.java
@@ -8,12 +8,12 @@ import java.util.List;
 public class JwtResponse {
     private String token;
     private String type = "Bearer";
-    private int id;
+    private Long id;
     private String username;
     private String email;
     private List<String> roles;
 
-    public JwtResponse(String accessToken, int id, String username, String email, List<String> roles) {
+    public JwtResponse(String accessToken, Long id, String username, String email, List<String> roles) {
         this.token = accessToken;
         this.id = id;
         this.username = username;
